apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.3
PyJWT==2.10.1
requests==2.32.3
azure-storage-blob==12.23.1
twilio==9.4.0
